# WebApi/Samples/api-swagger.yaml
openapi: 3.0.1
info:
  title: Web Game API
  version: '0.1'
paths:
  '/api/games/{gameId}':
    get:
      tags:
        - Games
      summary: Получить игру
      operationId: GetGameById
      parameters:
        - name: gameId
          in: path
          description: Идентификатор игры
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/GameDto'
        '404':
          description: Игра не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
  /api/games:
    post:
      tags:
        - Games
      summary: Создать игру
      operationId: CreateGame
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GameToCreateDto'
      responses:
        '201':
          description: Игра создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '400':
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '422':
          description: Ошибка при проверке
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
  '/api/games/{gameId}/players/{userId}':
    put:
      tags:
        - Games
      summary: Добавить игрока в игру
      description: "Имя игрока в каждой игре может быть уникальным, поэтому его надо передать.\r\nВ случае нескольких запросов для одного пользователя, должно быть установлено имя из последнего запроса."
      operationId: AddPlayerToGame
      parameters:
        - name: gameId
          in: path
          description: Идентификатор игры
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь добавлен в качестве игрока
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '400':
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '404':
          description: Игра или пользователь не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
    get:
      tags:
        - Games
      summary: Получить игрока в игре
      operationId: GetPlayerOfGame
      parameters:
        - name: gameId
          in: path
          description: Идентификатор игры
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/PlayerDto'
        '404':
          description: Игра или игрок не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
  '/api/games/{gameId}/status':
    get:
      tags:
        - Games
      summary: Получить статус игры
      operationId: GetGameStatusById
      parameters:
        - name: gameId
          in: path
          description: Идентификатор игры
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                enum:
                  - WaitingToStart
                  - Playing
                  - Finished
                  - Canceled
                type: string
            application/xml:
              schema:
                enum:
                  - WaitingToStart
                  - Playing
                  - Finished
                  - Canceled
                type: string
        '404':
          description: Игра не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
  '/api/games/{gameId}/players/{userId}/decision':
    post:
      tags:
        - Games
      summary: Задать решение игрока
      operationId: SetPlayerDecision
      parameters:
        - name: gameId
          in: path
          description: Идентификатор игры
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          description: Идентификатор пользователя
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              enum:
                - None
                - Rock
                - Scissors
                - Paper
              type: string
      responses:
        '200':
          description: ''
        '204':
          description: Решение задано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '400':
          description: Некорректные входные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
        '404':
          description: Игра или пользователь не найдены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Void'
            application/xml:
              schema:
                $ref: '#/components/schemas/Void'
...
components:
  schemas:
    GameDto:
      type: object
      properties:
        id:
          type: string
        players:
          uniqueItems: false
          type: array
          items:
            $ref: '#/components/schemas/PlayerDto'
        turnsCount:
          type: integer
          format: int32
        currentTurnIndex:
          type: integer
          format: int32
        status:
          enum:
            - WaitingToStart
            - Playing
            - Finished
            - Canceled
          type: string
    PlayerDto:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        decision:
          enum:
            - None
            - Rock
            - Scissors
            - Paper
          type: string
        score:
          type: integer
          format: int32
    Void:
      type: object
    GameToCreateDto:
      required:
        - turnsCount
      type: object
      properties:
        turnsCount:
          type: integer
          format: int32
...